requirements.txt
Flask==3.0.2
openai==1.14.3
ruamel.yaml==0.18.6